Use %d and (int) as the most portable.

+++ src/connection.c
@@ -568,8 +568,8 @@ int Connection_read_header(Connection *conn, Request *req)
 
         finished = Request_parse(req, conn->buf, conn->nread, &conn->nparsed);
 
-        check(finished != -1, "Error in parsing: %d, bytes: %d, value: %.*s, parsed: %lu", 
-                finished, conn->nread, conn->nread, conn->buf, conn->nparsed);
+        check(finished != -1, "Error in parsing: %d, bytes: %d, value: %.*s, parsed: %d", 
+                finished, conn->nread, conn->nread, conn->buf, (int)conn->nparsed);
 
     }
     check(finished, "HEADERS and/or request too big.");
